toaster: Increase maximum length of file path of build artifact
The default maximum length of the file path for a build artifact on
Krogoth Toaster is 100 characters (Django default for FilePathField).
This value makes it impossible to download build artifacts because
default file paths are quite long and therefore are being truncated
by Toaster.

Example of a long file path which does not work on my machine:
/var/www/toaster/poky/build-toaster-2/tmp/deploy/sdk/poky-glibc-x86_64-core-image-sato-cortexa8hf-neon-toolchain-2.1.1.sh

To fix this problem i increased the size of 'file_name' column
of the corresponding MySQL table to 255 characters.

[krogoth]

Signed-off-by: Alexander.Egorenkov@vector.com
Signed-off-by: Toaster <toasterbotldn@icarus>