toaster: buildinfohelper toaster-custom-images layer
This fixes the unidentified layers issue by making the
toaster-custom-images layer a local layer. By doing this we also fix the
git assumptions made for the local layers which stop recipes and other
meta data being associated with them. This also removed some of the
special casing previously needed when we didn't have the concept of a
local (non git) layer.

[YOCTO #10220]

Signed-off-by: Michael Wood <michael.g.wood@intel.com>
Signed-off-by: Toaster <toasterbotldn@icarus>