toaster: fix SDK artifact capture
Use the TaskArtifacts event to scan the SDK and ESDK manifests
to cleanly collect the respective artifact files.

The previous method was broken when the SDK file deployment moved
from the do_populate_sdk[_ext] tasks to their sstate tasks. That
method is disabled (but not yet removed) in preparation for the
rest of refactor work for the parent 10283 defect.

[YOCTO #10850]

Signed-off-by: David Reyna <David.Reyna@windriver.com>
Signed-off-by: Toaster <toasterbotldn@icarus>